Isolate source file memory utilization to Scanner
This change also remedies a memory leak from a dangling new in Lox::Run.

+++ src/scanner/token.cpp
@@ -4,8 +4,7 @@
 #include <string>
 
 std::ostream& operator<<(std::ostream& out, const Token& token) {
-  // return "<type_>" + ' ' + lexeme_ + ' ' + "<literal_>";
-  out << '[' << token.type_ << ' ' << token.lexeme_ << ' ' << token.line_
-      << ']';
+  out << '[' << static_cast<int>(token.type_) << ' ' << token.lexeme_ << ' '
+      << token.line_ << ']';
   return out;
 }
